refactor(folders): 👷 move 'project detail page' under layout
