+++ app/personal-projects/cryptomania/page.tsx
@@ -1,7 +1,7 @@
 import { Metadata } from 'next'
 
 import PageNavigation from '@/components/layout/page-navigation/PageNavigation'
-import ProjectPageLayout from '@/components/layout/projects/project-page/ProjectPageLayout'
+import ProjectPageLayout from '@/components/layout/projects/ProjectPageLayout'
 
 import { metaDataCryptomania } from '@/lib/data/metadata/pages/projects/personal/metaDataCryptomania'
 import { sections } from '@/lib/data/pages/projects/personal/cryptoMania'
