+++ components/layout/projects/project-page-detail/Gallery.tsx
