+++ components/layout/projects/project-page-detail/ProjectLinksList.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react'
 
-import GitHubRepositoryLink from '@/components/projects/project-detail-page/GitHubRepositoryLink'
+import GitHubRepositoryLink from '@/components/layout/projects/project-page-detail/GitHubRepositoryLink'
 import List from '@/components/shared/List'
 import ListItem from '@/components/shared/ListItem'
