+++ components/layout/projects/project-page-detail/ProjectResponsibility.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react'
 
-import PageSection from '@/components/layout/projects/project-page/PageSection'
+import PageSection from '@/components/layout/projects/project-page-detail/PageSection'
 
 import { Section } from '@/lib/utils/interfaces/interfaces'
