+++ components/layout/projects/project-page-detail/SkillsList.tsx
