+++ components/pages/projects/overview-page/ProjectImage.tsx
